backups/20241212_122405/requirements.txt
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
sentry-sdk>=1.35.0
structlog>=23.2.0
streamlit>=1.29.0
grpcio>=1.59.3
protobuf>=4.25.1
pyyaml>=6.0.1
psutil>=5.9.7
plotly>=5.18.0
pandas>=2.1.4
anthropic>=0.8.1
cryptography>=42.0.5
click>=8.1.7
mako>=1.3.0
pdoc3>=0.10.0
watchdog>=3.0.0
